requirements.txt
groq==1.0.0
PyMuPDF==1.26.7
python-dotenv==1.2.1
streamlit==1.53.0